# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SoundTrace API — Music identification, history, and trending
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface IdentifyInput {
  /** The video URL from Instagram, YouTube, Facebook, or any social platform */
  url: string;
}

export interface SongResult {
  id: number;
  title: string;
  artist: string;
  album: string;
  year: number;
  genre: string;
  spotifyUrl: string;
  youtubeUrl: string;
  spotifyId: string;
  youtubeId: string;
  /** @nullable */
  previewUrl?: string | null;
}

export interface HistoryEntry {
  id: number;
  title: string;
  artist: string;
  genre: string;
  searchedAt: string;
  spotifyUrl: string;
  youtubeUrl: string;
  spotifyId?: string;
  youtubeId?: string;
}

export interface TrendingEntry {
  rank: number;
  id: number;
  title: string;
  artist: string;
  genre: string;
  searchCount: number;
  growthPercent: number;
  isViral: boolean;
  spotifyUrl: string;
  youtubeUrl: string;
}

export interface TrendingResponse {
  songs: TrendingEntry[];
  isPremium: boolean;
  totalShown: number;
}

export interface AppStats {
  totalSearches: number;
  totalSongs: number;
  trendingCount: number;
  todaySearches: number;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface ErrorResponse {
  error: string;
}

export type GetTrendingParams = {
genre?: GetTrendingGenre;
region?: GetTrendingRegion;
};

export type GetTrendingGenre = typeof GetTrendingGenre[keyof typeof GetTrendingGenre];


export const GetTrendingGenre = {
  all: 'all',
  bollywood: 'bollywood',
  punjabi: 'punjabi',
  south: 'south',
  pop: 'pop',
  edm: 'edm',
} as const;

export type GetTrendingRegion = typeof GetTrendingRegion[keyof typeof GetTrendingRegion];


export const GetTrendingRegion = {
  all: 'all',
  north: 'north',
  south: 'south',
  west: 'west',
  east: 'east',
} as const;