# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SoundTrace API — Music identification, history, and trending
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /history/{id}:
    delete:
      operationId: deleteHistoryEntry
      tags: [history]
      summary: Delete a single history entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entry deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /songs/{id}:
    get:
      operationId: getSong
      tags: [songs]
      summary: Get song details by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Song details retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SongResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    SongResult:
      type: object
      required:
        - id
        - title
        - artist
        - album
        - year
        - genre
        - spotifyUrl
        - youtubeUrl
        - spotifyId
        - youtubeId
      properties:
        id:
          type: integer
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        year:
          type: integer
        genre:
          type: string
        spotifyUrl:
          type: string
        youtubeUrl:
          type: string
        spotifyId:
          type: string
        youtubeId:
          type: string
        previewUrl:
          type: ["string", "null"]
...
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string