# lib/api-zod/src/generated/types/songResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SoundTrace API — Music identification, history, and trending
 * OpenAPI spec version: 0.1.0
 */

export interface SongResult {
  id: number;
  title: string;
  artist: string;
  album: string;
  year: number;
  genre: string;
  spotifyUrl: string;
  youtubeUrl: string;
  spotifyId: string;
  youtubeId: string;
  /** @nullable */
  previewUrl?: string | null;
}